refactor: 모달창 컴포넌트화
